// admin/types/type.d.ts
export interface BilingualText {
  en: string;
  vi: string;
}

export interface Phonetic {
  text: string;
  audio: string;
  locale: string;
}

export interface Category {
  category_id: string;
  name: string;
  description: string;
  image_url: string;
  created_at: string;
}

export interface CategoryOption {
  id: string;
  name: string;
  description: string;
}

export interface Card {
  card_id: string;
  word: string;
  type: string;
  explanation: BilingualText;
  translation: string;
  example: BilingualText;
  phonetics: Phonetic[];
  image_url: string;
  difficulty: string;
  category_id: string;
}

export type CardForm = Omit<Card, "card_id">;